+++ src/interfaces/index.ts
@@ -1,4 +1,5 @@
 export * from './initState/initState.interface';
+export * from './keyboardAction/keyboardAction.interface';
 export * from './normalStateContext/normalStateContext.interface';
 export * from './optionsGatherer/optionsGatherer.interface';
 export * from './pluginDescription/pluginDescription.interface';