+++ src/plugins/interactions/simpleInteractions/simpleInteractions.component.ts
@@ -1,5 +1,5 @@
-import {ChangeDetectionStrategy, Component, ElementRef, Inject, inject, OnDestroy, Optional} from '@angular/core';
-import {RecursivePartial} from '@jscrpt/common';
+import {ChangeDetectionStrategy, Component, DOCUMENT, effect, ElementRef, Inject, inject, OnDestroy, Optional} from '@angular/core';
+import {NoopAction, RecursivePartial} from '@jscrpt/common';
 import {deepCopyWithArrayOverride} from '@jscrpt/common/lodash';
 import {Subscription} from 'rxjs';
 
@@ -7,7 +7,7 @@ import {Interactions, InteractionsOptions} from '../../../interfaces';
 import {SelectPluginInstances, SelectBus} from '../../../misc/classes';
 import {CopyOptionsAsSignal} from '../../../decorators';
 import {INTERACTIONS_OPTIONS} from '../../../misc/tokens';
-import {handleOptionClick, togglePopupOnClick} from '../../../misc/utils';
+import {handleClickOutside, handleOptionClick, togglePopupOnClick} from '../../../misc/utils';
 
 const defaultOptions: InteractionsOptions =
 {
@@ -36,6 +36,16 @@ export class SimpleInteractions<TValue = unknown> implements Interactions<TValue
      */
     protected optionClick: Subscription = new Subscription();
 
+    /**
+     * Instance of HTML document
+     */
+    protected document: Document = inject(DOCUMENT);
+
+    /**
+     * Method used for unregistering click outside handler
+     */
+    protected clickOutsideUnregister: NoopAction|undefined|null;
+
     //######################### public properties - implementation of SelectPlugin #########################
 
     /**
@@ -57,7 +67,7 @@ export class SimpleInteractions<TValue = unknown> implements Interactions<TValue
     /**
      * @inheritdoc
      */
-    public selectBus: SelectBus<TValue> = inject(SelectBus);
+    public selectBus: SelectBus<TValue> = inject(SelectBus) as SelectBus<TValue>;
 
     //######################### constructor #########################
     constructor(@Inject(INTERACTIONS_OPTIONS) @Optional() options?: RecursivePartial<InteractionsOptions>|null,)
@@ -67,6 +77,22 @@ export class SimpleInteractions<TValue = unknown> implements Interactions<TValue
 
         togglePopupOnClick(this.selectBus, this.clickPopupVisibleToggle);
         handleOptionClick(this.selectBus, this.selectPlugins, this.optionClick);
+
+        effect(() =>
+        {
+            const popupVisible = this.selectBus.popupVisible();
+
+            this.clickOutsideUnregister?.();
+
+            if(popupVisible)
+            {
+                this.clickOutsideUnregister = handleClickOutside(this.document, this.selectBus, this.selectPlugins);
+            }
+            else
+            {
+                this.clickOutsideUnregister = null;
+            }
+        });
     }
 
     //######################### public methods - implementation of SelectPlugin #########################