+++ src/plugins/positioner/noPositioner/noPositioner.component.ts
@@ -36,7 +36,7 @@ export class NoPositioner<TValue = unknown> implements Positioner<TValue, Positi
     /**
      * @inheritdoc
      */
-    public selectBus: SelectBus<TValue> = inject(SelectBus);
+    public selectBus: SelectBus<TValue> = inject(SelectBus) as SelectBus<TValue>;
 
     //######################### public methods - implementation of SelectPlugin #########################