+++ src/plugins/valueHandler/staticValueHandler/staticValueHandler.component.ts
@@ -51,7 +51,7 @@ export class StaticValueHandler<TValue = unknown> implements ValueHandler<TValue
     /**
      * @inheritdoc
      */
-    public selectBus: SelectBus<TValue> = inject(SelectBus);
+    public selectBus: SelectBus<TValue> = inject(SelectBus) as SelectBus<TValue>;
 
     //######################### public properties - implementation of ValueHandler #########################